Convert Insider list to yaml data (#14)
Make working with YAML files easier in vscode.

The insider list is now a data-driven YAML file. Much of the old CMS's markup was
removed as well. A "fresh" css file was added so that we do not continue to contribute to the CMS's stylesheet. Eventually, the "fresh" stylesheet will replace the current one.